Add check for errors when launching files with associations
