+++ spyder/plugins/explorer/widgets/fileassociations.py
@@ -25,10 +25,10 @@ from qtpy.QtWidgets import (QApplication, QDialog, QDialogButtonBox,
 # Local imports
 from spyder.config.base import _
 from spyder.plugins.explorer.utils import (get_application_icon,
-                                           get_installed_applications,
-                                           parse_linux_desktop_entry)
+                                           get_installed_applications)
 from spyder.utils import icon_manager as ima
 from spyder.utils.encoding import is_text_file
+from spyder.utils.programs import parse_linux_desktop_entry
 
 
 class InputTextDialog(QDialog):
